refactor: infrastructure層の実体をshared/servicesに移動（TASK-0102）
DIコンテナ、リポジトリ、ローダーの実体をshared/services/配下に移動し、
旧infrastructure/ファイルを後方互換性用の再エクスポートに変換。

移動内容:
- infrastructure/di/ → shared/services/di/
- infrastructure/repositories/ → shared/services/repositories/
- infrastructure/loaders/ → shared/services/loaders/
