+++ atelier-guild-rank/src/infrastructure/di/index.ts
@@ -1,13 +1,15 @@
 /**
- * Infrastructure DI Container
- * DIコンテナの公開エクスポート
+ * Infrastructure DI Container（後方互換性用再エクスポート）
  *
- * TASK-0028: サービス統合・DI設定
+ * @description
+ * 実体は shared/services/di/ に移動済み。
+ * 後方互換性のため再エクスポートを提供する。
+ *
+ * 新規コードでは @shared/services を使用すること。
  */
-
-export { Container, type ServiceKey, ServiceKeys } from './container';
+export { Container, type ServiceKey, ServiceKeys } from '@shared/services/di/container';
 export {
   initializeServices,
   resetServices,
   type ServiceInitializationConfig,
-} from './setup';
+} from '@shared/services/di/setup';
