+++ atelier-guild-rank/src/shared/services/di/index.ts
@@ -1,0 +1,13 @@
+/**
+ * Shared Services DI Container
+ * DIコンテナの公開エクスポート
+ *
+ * TASK-0102: infrastructure層からshared/services/diに移動
+ */
+
+export { Container, type ServiceKey, ServiceKeys } from './container';
+export {
+  initializeServices,
+  resetServices,
+  type ServiceInitializationConfig,
+} from './setup';
