+++ atelier-guild-rank/src/shared/services/loaders/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Shared Services Loaders
+ * データローダーの公開エクスポート
+ *
+ * TASK-0102: infrastructure層からshared/services/loadersに移動
+ */
+
+export type { IJsonLoader } from './json-loader';
+export { JsonLoader } from './json-loader';
